+++ pandapower/control/__init__.py
@@ -9,6 +9,7 @@ from pandapower.control.controller.shunt_control import DiscreteShuntController
 from pandapower.control.controller.station_control import BinarySearchControl, DroopControl, VDroopControl_local
 from pandapower.control.controller.trafo.ContinuousTapControl import ContinuousTapControl
 from pandapower.control.controller.trafo.DiscreteTapControl import DiscreteTapControl
+from pandapower.control.controller.trafo.LineDropCompensationTapControl import LineDropCompensationTapControl
 from pandapower.control.controller.trafo.TapDependentImpedance import TapDependentImpedance
 from pandapower.control.controller.trafo.VmSetTapControl import VmSetTapControl
 from pandapower.control.controller.trafo_control import TrafoController